Handle quoted multipart/form-data boundary value.
I don't expect a Web browser to ever quote the boundary value
as they can just stick to a simple character set. But clients
may behave differently when performing REST calls for example.

No signs of quotes in grammar of the old RFC1521. But RFC2046
says "it does not hurt" and RFC7578 (referenced by the HTML 5
W3C standard) emphasises that "it is often necessary".

Ran into this scenario when using QtNetwork's QHttpMultiPart
class for a command line client.