Add config for database to execute queries on open
With the ConnectOptions setting in database.ini, you can already specify
QSqlDatabase::setConnectOptions(). But enabling foreign key constraints
for SQLite requires executing the SQL query

    PRAGMA foreign_keys = ON;

for each QSqlDatabase connection after it was opened. This is not
possible with the ConnectOptions setting and there is no other way to
achieve this.

With this change, you can now add

    OnDatabaseOpen/size=1
    OnDatabaseOpen/1/query=PRAGMA foreign_keys = ON;

to database.ini to enable foreign key support.

I chose the generic approach to allow specifying an array of queries
since with QSqlQuery/TSqlQuery, only one SQL statement can be executed
at once, so with this approach, it is easy to specify more SQL queries
if that is ever needed.