Support SMTP connections to older servers not supporting ESMTP yet
This adjusts the initial SMTP handshake performed by Teamserver such
that in case EHLO fails, it'll fall back to HELO (and also avoid using
other ESMTP commands such as STARTTLS or AUTH). This is per the
recommendation in RFC 1869 which says that

> In the case of any error response, the client SMTP should issue either
the HELO or QUIT command.

This enables Teamserver to talk to older or very primitive SMTP servers
not supporting ESMTP. In particular, it allows using the
'DebuggingServer' class which is part of the 'smtp' module shipped with
Python.